backend/docker_vendor/growwapi/pyproject.toml
[tool.poetry]
name = "growwapi"
version = "1.5.0"
description = "The foundational SDK for accessing Groww APIs and listening to live data streams. This package provides the core functionalities required to interact with Groww's trading platform."
authors = ["Groww <growwapi@groww.in>"]
readme = "README.md"
packages = [
    { include = "growwapi", from = ".." },
]
license = "MIT"
[tool.poetry.dependencies]
aiohttp = "^3.11.18"
clint = "^0.5.1"
google = "^3.0.0"
nats-py = "^2.9.0"
nkeys = "^0.2.1"
protobuf = "^5.29.0"
pydantic = "^2.10.2"
python = "^3.9"
requests = "^2.32.3"
pynacl = "^1.5.0"
pandas = "^2.2.3"
certifi = "^2025.1.31"
colorama = "^0.4.6"

[tool.poetry.group.dev.dependencies]
pyright = "^1.1.383"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
